requirements.txt
slack_bolt
slack_sdk
openai
pandas
openpyxl
python-dotenv
requests
